+++ R/GetSMNData.R
@@ -1,55 +1,76 @@
 #' Get Meteorological data
 #'
 #' Downloads minimum and maximum temperature station data from
-#' Argentina's National Weather Service's
-#' public access. Data availability is restricted to one year prior.
+#' Argentina's National Weather Service's public access. Data availability is not
+#' guaranteed so you are encouraged to check it on
+#' [the website](https://www.smn.gob.ar/descarga-de-datos).
 #'
 #' @param date date vector of dates to fetch data
+#' @param type type of data to retrieve
 #' @param bar logical object indicating whether to show a progress bar
 #'
 #' @return
-#' a dataframe with daily minimum and maximum temperature for all available stations
-#' and each date.
+#' For `type = "hourly"`, a data.frame with observations of
+#' \describe{
+#'    \item{date}{date}
+#'    \item{t}{temperature in degrees celsius}
+#'    \item{rh}{relative humidity in \%}
+#'    \item{slp}{sea level pressure in hPa}
+#'    \item{dir}{wind direction in clockwise degrees from 6 o'clock}
+#'    \item{V}{wind magnitude in m/s}
+#'    \item{station}{station name}
+#' }
+#'
+#' For `type = "daily"`, a data.frame with observations of
+#' \describe{
+#'    \item{date}{date}
+#'    \item{tmax}{maximum daily temperature in degrees celsius}
+#'    \item{tmin}{minimum daily temperature in degrees celsius}
+#'    \item{station}{station name}
+#' }
+#'
+#' For `type = "radiation"`, a data.frame with observations of
+#' \describe{
+#'    \item{date}{date}
+#'    \item{global}{global radiation in W/m^2}
+#'    \item{diffuse}{diffuse radiation in W/m^2}
+#'    \item{station}{station name}
+#' }
+#'
 #'
 #' @examples
 #' \dontrun{
 #' dates <- seq.Date(lubridate::today() - 30, lubridate::today(), by = "1 day")
-#' data <- GetSMNData(dates, bar = FALSE)    # long!
+#' data <- GetSMNData(dates, type = "daily", bar = TRUE)
 #'
 #' library(ggplot2)
 #' ggplot(subset(data, station == "BASE BELGRANO II"),
-#'        aes(date, (t.max + t.min)/2)) +
+#'        aes(date, (tmax + tmin)/2)) +
 #'     geom_line()
 #' }
 #'
 #' @source https://ssl.smn.gob.ar/dpd/pron5d-calendario.php
 #' @export
 #' @import RCurl
-GetSMNData <- function(date, type = c("hourly", "daily", "current", "radiation"),  bar = FALSE) {
-    return.data <- data.frame()
+GetSMNData <- function(date, type = c("hourly", "daily", "radiation"),  bar = FALSE) {
     no_data <- vector()
 
-    if (bar == TRUE) pb <- txtProgressBar(min = 1, max = length(date), style = 3)
+    if (bar == TRUE) pb <- txtProgressBar(min = 0, max = length(date), style = 3)
 
-    for (i in seq_along(date)) {
+    return.data <- rbindlist(lapply(seq_along(date), function(i) {
         if (bar == TRUE) setTxtProgressBar(pb, i)
-
-        base.url <- "https://ssl.smn.gob.ar/dpd/descarga_pron5d.php?file="
-        file.url <- paste0("obs", gsub("-", "", date[i]), ".txt")
-        file.url <- RCurl::base64Encode(file.url)[1]
-        file.url <- RCurl::base64Encode(file.url)[1]
-        url <- paste0(base.url, file.url)
-
-        obs <- readLines(url, warn = F)
-
-        if (obs[1] != "El archivo no existe.") {
-            data <- .CleanSMN(obs)
-            data$date <- as.Date(date[i])
-            return.data <- rbind(return.data, data)
+        if (type[1] == "hourly") {
+            data <- .smnhourly(date[i])
+        } else if (type[1] == "daily") {
+            data <- .smnobs(date[i])
+        } else if (type[1] == "radiation") {
+            data <- .smnrad(date[i])
         } else {
-            no_data <- c(no_data, i)
+            stop("wrong type", .call = FALSE)
         }
-    }
+        if (is.null(data)) no_data <- i
+        return(data)
+    }))
 
     if (bar == TRUE) close(pb)
 
@@ -60,21 +81,9 @@ GetSMNData <- function(date, type = c("hourly", "daily", "current", "radiation")
     return(return.data)
 }
 
-
-.CleanSMN <- function(obs) {
-    obs <- obs[4:length(obs)]
-    t.max <- as.numeric(stringr::str_sub(obs, 10, 15))
-    t.min <- as.numeric(stringr::str_sub(obs, 16, 21))
-    station <- as.character(stringr::str_sub(obs, 22))
-    station <- stringr::str_trim(station)
-    return(data.frame(t.max, t.min, station, stringsAsFactors = F))
-}
-
-
-
 .smnhourly <- function(date) {
     file <- paste0("observaciones/datohorario",
-                   stringr::str_remove_all(as.character(as.Date(date)), "-"), ".txt")
+                   stringr::str_remove_all(as.character(date), "-"), ".txt")
     url <- paste0("https://ssl.smn.gob.ar/dpd/descarga_opendata.php?file=", file)
     obs <- readLines(url, warn = F)
     if (obs[1] != paste0(file, "El archivo no existe.")) {
@@ -96,7 +105,7 @@ GetSMNData <- function(date, type = c("hourly", "daily", "current", "radiation")
         lubridate::hour(date) <- obs$hour
         obs$date <- lubridate::as_datetime(date)
         obs[, hour := NULL]
-        obs
+        return(obs)
     } else {
         return(NULL)
     }
@@ -124,7 +133,7 @@ GetSMNData <- function(date, type = c("hourly", "daily", "current", "radiation")
         obs <- obs[!is.na(date)]
 
         obs$date <- date
-        obs
+        return(obs)
     } else {
         return(NULL)
     }
@@ -140,13 +149,13 @@ GetSMNData <- function(date, type = c("hourly", "daily", "current", "radiation")
 
     bs <- obs[, 1:3]
     setnames(bs, c("date", "global", "diffuse"))
-    bs[, station := "BsAs"]
+    bs$station <- "BsAs"
 
     ush <- obs[, c(1, 4:5)]
     setnames(ush, c("date", "global", "diffuse"))
-    ush[, station := "Ushuaia"]
+    ush$staton <- "Ushuaia"
 
     obs <- rbind(bs, ush)
     obs[, date := lubridate::as_datetime(date)]
-    obs
+    return(obs)
 }